+++ test/pallet/core/jclouds_core_test.clj
@@ -1,13 +1,10 @@
 (ns pallet.core.jclouds-core-test)
 
 (try
-  (use '[pallet.core.api :only [version]])
-  (catch Throwable _ ; java.lang.IllegalAccessError
-    (try
-      (use '[pallet.core :only [version]])
-      (catch Throwable _
-        (defn version [] "0.8.0-alpha.2-or-before")))))
+  (require '[pallet.api :refer [version]])
+  (catch Throwable e ; java.lang.IllegalAccessError
+    (require '[pallet.core :refer [version]])))
 
-(if (.startsWith (version) "0.7")
+(if (and (string? (version)) (.startsWith (version) "0.7"))
   (load "jclouds_core_test_0_7")
   (load "jclouds_core_test_0_8"))